modularized parts till binding adapters which are problematic
