+++ .idea/gradle.xml
@@ -12,10 +12,16 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/functionality" />
+            <option value="$PROJECT_DIR$/functionality/detail" />
+            <option value="$PROJECT_DIR$/functionality/detail/entities" />
             <option value="$PROJECT_DIR$/functionality/search" />
             <option value="$PROJECT_DIR$/functionality/search/entities" />
             <option value="$PROJECT_DIR$/functionality/shared" />
             <option value="$PROJECT_DIR$/functionality/shared/met-api" />
+            <option value="$PROJECT_DIR$/ui" />
+            <option value="$PROJECT_DIR$/ui/common" />
+            <option value="$PROJECT_DIR$/ui/detail" />
+            <option value="$PROJECT_DIR$/ui/search" />
           </set>
         </option>
       </GradleProjectSettings>
