--- app/src/main/java/com/example/metmuseum/detail/DetailUiModel.kt
@@ -1,17 +1,0 @@
-package com.example.metmuseum.detail
-
-import com.example.met_api.model.MetPhoto
-
-/**
- * Detail class that gets the mapped data from the APIs result.
- */
-data class DetailUiModel(
-    val imgUrl: String,
-    val title: String,
-    val objectDate: String,
-    val department: String,
-    val repository: String,
-    val artist: String,
-    val culture: String,
-    val imageList: List<com.example.met_api.model.MetPhoto>
-)
