+++ settings.gradle
@@ -19,3 +19,9 @@ include ':functionality:shared'
 include ':functionality:shared:met-api'
 include ':functionality:search'
 include ':functionality:search:entities'
+include ':functionality:detail'
+include ':functionality:detail:entities'
+include ':ui'
+include ':ui:search'
+include ':ui:detail'
+include ':ui:common'
